gui_dialogs: Fix crash when filename doesn't exist
This fixes a crash in the file open dialog when a manually entered
filename does not exist. This is a minor change, as opening a file which
does not exist is not very useful. However, this dialog code will likely
be reused in the future as a file save dialog, where this functionality
will be very important.

The change is made by switching from 'fs::canonical' to 'fs::absolute',
as 'canonical' requires the full path to be present on the filesystem,
where as 'absolute' does not. There are two important consequences of
this change:

 1. The file browser will no longer actively follow symlinks. This does
 not mean the user cannot browse into symlinks, rather it means the path
 represented to the user in the browser (and also returned by the dialog
 itself) will continue to show the symlink path rather than the real
 path referred to by the symlinks.

 2. The file browser will no longer normalize path names, specifically
 path elements such as "." or ".." will not be resolved out of the path.
 For example, if a user enters the path "/opt/foo/../bar", this will no
 longer be reduced to "/opt/bar" as it had been before. This also means
 that if a user has browsed to "/opt/foo/../bar" and clicks the up
 button (labeled "<-"), they will browse to "/opt/foo/.." which will
 show the contents of "/opt". Clicking up again would then browse to
 "/opt/foo".

Consequence 1 may be considered a positive design change, and may or may
not be reverted in the future.

Consequence 2 is a potentially serious issue for usability and should
be fixed later on. This would have been done using 'lexically_normal',
however this function is not available under the experimental FS
library.