nsf: Fix item/page vector construction bug
Prior to this change, certain vectors in the NSF code were constructed
like so:

  std::vector<int> foo = {
      &bar[i],
      &bar[i + n]
  };

Where 'bar.size()' is potentially 'i + n'. This is undefined behavior as
the second usage of 'operator []' on 'bar' would result in dereferencing
a pointer at the end of the vector storage for 'bar'. Such usage trips
an assertion in the MSVC STL implementation.

The change alters the above statement to become:

  std::vector<int> foo = {
      bar.data() + i,
      bar.data() + i + n
  };

...which should be acceptable.