Fix fstream usage on Windows
Prior to this change, usage of fstream in DRNSF operates on two
assumptions not valid on some Windows implementations:

 1. There is no difference between text mode and binary mode when
 opening a file for reading or writing.

 2. All filenames may be represented to 'fstream::open' or the 'fstream'
 constructor in the form of a narrow string ('char[]', 'std::string').

As of this change, binary files are now opened using the function
'util::fstream_open_bin'. This function includes the binary file mode
flag, and on Windows converts the filename (as UTF-8) to a wide string.