Add 'STATIC_MSVC_RUNTIME' build option
By default, MSVC appears to link against the shared DLL versions of the
MSVC runtime. This behavior may be undesirable for some users, as the
resulting executable will depend on installation of the Visual C++
Redistributable Package for whatever version of MSVC was used when
building DRNSF.

This change adds a CMake build option which, when enabled, changes
compiler flags to link against the appropriate MSVC static runtime. This
option is disabled by default, preserving the previous default behavior.