render_world_fig: Fix bug with preexisting worlds
When a world_fig becomes associated with a world asset, the position of
that world_fig within the 3D scene should be immediately updated to
match the world's position. Due to a programming mistake, this does not
occur. In the current application, this manifests as a user-obserable
bug through the following method:

 1. Open an NSF file containing some scenery not located at (0, 0, 0).

 2. Delete at least one such scenery not located at (0, 0, 0).

 3. Undo the deletion.

 4. The scenery appears at (0, 0, 0).

The initial import does not display any issue because part of the import
process involves first creating the world asset and then setting its
world position. Setting the world position raises an event which the
associated world_fig *does* recognize and so sets its own position
within the viewport. However, when undoing a deletion, the world comes
back into existence with its position already initialized, so no change
event is raised, and the newly associated world_fig keeps its initial
position.