gui: Implement WM_DELETE_WINDOW for X11
When the user attempts to close one of the application's top level
windows via the window manager, a new 'on_close_request' function is
called on the window in question. Prior to this change, the behavior
of the window manager would likely be to terminate the application's
X connection, resulting in the process's ungraceful exit.