Swap Y and Z coordinates for all 3D code
This change establishes a new standard for the orientation of the 3D
world space in DRNSF. Z values increase upwards, Y values increase
towards the viewer, and X values increase towards the right. This is a
departure from the previous (CrashEdit-compatible) standard which has Y
and Z swapped.

Across the various model and world formats throughout the three games,
there is no fully consistent ordering to the coordinate values stored in
each entry. For this reason, a standard was picked based on typical
coordinate systems from existing map editors for other game engines.

As a consequence of the axis swap, the green and blue markers in the 3D
viewer reticle have swapped as well, as they continue to match X/Y/Z
with R/G/B.