Initial work on new property editor
This new property editor utilizes a new reflection-like system for
inspecting asset types to enumerate their properties and provide a UI
element for each property to allow the user to edit the property values.
This is implemented using a new template type `edit::field<T>', which
attempts to provide an interface for editing a value of type `T'. This
currently supports the following types:

 * Integral types (int, short, long, etc; currently read-only)
 * Floating types (float, double; currently read-only)
 * Vectors of any type