core: Add 'cdxa-imprint' command
This command is primarily intended for branding a BIN file with a
particular region license. The license includes some information such as
"Licensed by Sony Computer Entertainment" as well as the PlayStation
logo which appears when booting the game. The PlayStation system will
not boot a game disc which does not include this exact data (although
many workarounds often bypass this check regardless). Some emulators
also require this file (mednafen), or use it to autodetect the region.

The PlayStation license scheme is a trick to require unlicensed games to
include the "Licensed by Sony" text and (possibly trademarked) PS logo,
potentially opening up the publishers to lawsuits. In the United States,
prior case law (Sega v. Accolade) covers a similar scenario.