res: Add a res::tracker clone for entire subtrees
This new type, `res::tree_tracker<T>' operates similarly to
`res::tracker<T>', but tracks a full tree of asset names. Usually, this
will be used to track an entire project, but it could be used in the
future to track subtrees. This may be useful if multiple game levels are
loaded into a single project under different subtrees.