gui: Replace file dialog functions with a class
This change replaces the `show_open_dialog' and `show_save_dialog'
functions with a single `file_dialog' class with `run_open' and
`run_save' methods. The purpose of this class is to preserve some dialog
state (such as the current directory) between invocations, as well as to
simplify configuration of advanced options in the future (filters,
readonly checkbox, etc).

The implementation of the `file_dialog' is concealed in an impl class
following the common C++ PIMPL idiom. This is done to allow the dialog
to be implemented differently for some platforms or frontends (WINAPI)
where there may be a native implementation available.