gui: Fix file dialog issues with UTF-8 on Windows
This change fixes a bug where filenames with non-ASCII characters would
be improperly treated by the file browser, resulting in crashes when
attempting to select files with such names or when manually entering
such a name.

This does not affect other platforms where narrow strings are assumed to
be UTF-8.