render::viewport: Fix jumping bug on first input
This change fixes a bug where the first input given to a render viewport
after some time is combined with a (potentially large) built up delta
time from the previous update() call. Normally, these calls are very
rapid due to ImGui widgets operating with a high update frequency, so
the jump is small but still perceptible. In the future, this may not be
the case, and delays could build up to 4 seconds.