gui: Move widget updates to new 'core::worker'
This change removes the 'widget::update' method entirely, replacing the
functionality with a new 'core::worker' abstract class. The workers are
separate from the GUI code, allowing non-GUI code to install code which
should be called routinely or at a particular interval. This also
reduces the number of objects which need to be touched during updates,
as many widgets do not require them.