edit::asset_propctl: Reduce label column width
Prior to this change, the label column and field/editor column in the
asset property editor shared equal width. This change forces the label
column to a width of approximately 24 characters of text, providing more
space to the field editors and allowing the user to grow the window size
to expand the field column without the label column stealing 50% of the
gains.