nsf::eid: Add invalid-EID markers to EID strings
Prior to this change, when an EID is converted to a string, the state of
the high bit and low bit are not represented in that string. This is not
an issue for most EID's, as the two bits have defined values. However,
in the case of an invalid EID, this invalid-bit state is discarded when
converting to string, and would not be preserved when converting back
from string to EID.

This change prepends and/or appends an equal sign to the string when the
EID's high and/or low bit (respectively) is invalid, thus preserving the
invalid part of the state.