gui: Fix numpad keycodes on WINAPI
Prior to this change, numpad keys would yield non-numlock keycodes such
as home, end, pageup/down, etc even if numlock was enabled. This was due
to passing the WM_KEYDOWN and WM_KEYUP scancodes through MapVirtualKey
instead of taking the virtual keycode directly from the message.

MapVirtualKey is still used for ctrl, shift, and alt keys as this is
necessary to differentiate between left-hand keys and right-hand keys.

This change fixes a bug in the 'edit::field' hex editor where typing on
the numpad would also scroll around the view as if home/end/etc were
being pressed.