edit: Place NSF entries into their own subtree
This changes the NSF loading code so that pagelets are renamed based on
their EID (e.g. "entries/50_eZ") rather than remaining with the previous
pagelet name (e.g. "nsfile/page-12/pagelet-4"). This, combined with the
sorted asset treeview, provides a much faster and simpler way for users
to locate and identify entry assets in the editor.