gui::run: Fix X11 queue blocking issues
This fixes a bug where 'select()' would be called with an unintendedly
large timeout when processing multiple batches of messages in sequence.

This bug appears as a completely blank UI at first, with the menu bar
items not drawing until the window is moved, resized, or the mouse comes
across the menu bar.

The unintended program flow was as such:

 1. Workers are updated, yielding a large timeout.

 2. 'XQLength' returns a non-zero value. The select code is skipped,
 however the timeout-zeroing code is also skipped. The timeout continues
 to hold a large value.

 3. Pending X events are processed. The event processing block continues
 executing from the beginning of the loop. A comment indicates that
 'select' will be called with a zero timeout, however the timeout is not
 zero because the previous block was skipped.

 4. 'select' is called with a large timeout. Some widgets are marked as
 dirty and are ready to be redrawn, but the redrawing code will not be
 reached until the timeout is met (many years in the future) or an event
 arrives to answer the select call.

Previously, this was not an issue as ImGui widgets provide a low
timeout. Now with the new editor mode system, the initial UI display
does not include any such widgets, and thus a high timeout is obtained,
exposing the bug.