edit::asset_tree: Switch to using tree_tracker
This change fixes an issue where any assets appearing prior to the
construction of an asset tree view ('edit::asset_tree') do not appear in
that tree view until they disappear and reappear. An example flow
demonstrating this issue:

 1. The user starts the application with a blank interface. No asset
 tree appears.

 2. The user opens an NSF file. Numerous assets are created as part of
 this process.

 3. The user switches to an editor mode containing an asset tree view.
 The mode has not been previously selected during this session, so the
 mode's UI (including the asset tree) is constructed on demand. The tree
 appears blank despite the user having previously loaded an NSF file.

 4. The user clicks "Undo" in the edit menu. The assets vanish from the
 project, but no change appears in the still-empty asset tree.

 5. The user clicks "Redo" in the edit menu. The assets reappear in the
 project, and the asset tree view reacts by populating itself with tree
 nodes representing the assets.

This change replaces the existing system which directly monitors a
project by-reference with the tree_tracker already used by the map mode.
The tree tracker does not exhibit this issue, as the initial setting of
a base name for tracking raises events for the pre-existing assets. As a
side benefit, the asset tree is now capable of switching between
projects, or even switching to display a limited subtree of a project's
asset name space.