edit::asset_tree: Sort tree nodes by name
Prior to this change, asset trees were not meaningfully ordered. This
change orders the asset tree nodes using a simple less-than comparison
of the strings for each asset name.

Also prior to this change, tree nodes in the general 'gui::treeview'
widget were not given a documented order. This change specifies an
insert-at-end or insert-at-partition-by-predicate behavior, which
differs from the previous insert-at-front behavior which was not
documented or particularly desirable.

Note that 'std::find_if_not' is used over 'std::partition_point' as the
latter requires the node list to be properly partitioned, which is not a
strict requirement presented by the documentation for the new node
constructor code. This may be improved in the future.

As an additional change, the treeview node and subnode lists are
replaced with vector types which may yield a meaningful performance
increase.