edit: Add "assets" mode
This mode is identical to the existing interrim mode which provides a
tree of selectable assets on the left-hand side for the user to select
and displays information, options, and editable properties of the
selected asset on the right-hand side.

Unlike the existing interrim mode, this assets mode supports switching
between projects. This is necessary to more properly implement
New/Open/Close functionality.