edit: Make 'Open' create a new project
Prior to this change, an editor context would be constructed with an
existing (blank) project and the "Open" menu item would simply import
NSF-related assets into the existing project. This commit changes the
behavior such that the context is now constructed without a project at
all, and the "Open" menu item creates a fresh project to import the NSF
data into each time.

This also fixes a bug where the user could not use "Open" twice during
one program execution without manually undoing the import enacted by the
previous one. This occurred because the second import would attempt to
construct assets with names already in use, such as 'nsfile'.