gl: Handle and log GL program link errors
Prior to this change, various calls are made throughout the program to
'glLinkProgram' without checking for any link failures. This change
replaces those calls with calls to a new 'gl::link_program' function
which checks for link errors and logs them in addition to throwing an
exception which captures some GL diagnostic information.