gui::run: Fix WINAPI queue blocking issues
This change fixes a bug similar to the one addressed by commit
b09a0425118ac250ab30ef53c3aa27f084c12494.

In this case, 'MsgWaitForMultipleObjects' is called with a large timeout
but blocks despite the presence of pending window messages in the thread
message queue. This occurs because of a prior call to 'PeekMessage' (or
related functions from a modal message loop). MsgWaitForMultipleObjects
considers all pending messages queued prior to a call to PeekMessage/etc
to be ineligible for unblocking.

Because of this, the user interface can appear sluggish and glitchy, as
deferred WM_PAINT messages are often left unprocessed before the thread
blocks. When the thread unblocks from a non-WM_PAINT message, the
WM_PAINT messages may be further deferred as the thread blocks again.

The fix is to replace this with a call to 'MsgWaitForMultipleObjectsEx'
with the flag 'MWMO_INPUTAVAILABLE' which does consider those messages.