Implement native window menubars for WINAPI
A new build option, "FEATURE_NATIVE_MENU", has been added to control
whether or not to use these native menus. By default, it is enabled.

This change also introduces a 'gui::command' type which allows objects
to allocate a global ID which can be used to receive WM_COMMAND
messages.