gl: Separate source specification from compilation
Prior to this commit, the function 'gl::compile_shader' takes a single
shader source string and both attaches the source to the shader as well
as compiles the shader and checks for errors.

This change separates the source specification into a new function,
'gl::shader_source'. Unlike the old 'gl::compile_shader' call, this new
function can take multiple sources at once.

The new function takes 'string_view' objects, so it can support a wider
variety of string formats copy-free. The 'DRNSF_DECLARE_EMBED' macro has
been expanded to provide an 'str' member of this type to simplify the
source specification for shaders.