Add a temporary replacement for std::string_view
Due to changes in a previous commit, the program no longer builds on GCC
6.3.0, as used in the current Debian Stable (Stretch) and tested under
one of the configurations with Travis CI.

This change replaces string_view with a temporary, less-featureful
replacement. This change should be rolled back in the future when the
build requirements are bumped up to GCC 7 which does successfully build
the program.