nsf::zdat_v2: Add initial (broken) property parser
This code provides the basis for a Crash 2 object and camera property
parser. Each propertybag-like item is loaded into a 'game::attr_table'
object.

This code was written with the false assumption that each value group
within a columned row has a unique column id. This is not the case for
some camera properties in some levels, including the following property
numbers:

 * 0x119
 * 0x142
 * 0x16E
 * 0x183
 * 0x185
 * 0x186
 * 0x1B5
 * 0x1F9
 * 0x1FA
 * 0x297

Some additional properties also encounter this issue in Crash 3. A list
of those property numbers was not compiled.