Fix GL viewport queue bloat issues on X11
This change introduces a new `ready_to_draw' virtual method for all
non-composite widgets. The default behavior is to always return true.
This method has been overridden in render::viewport to return false if
the GPU has not finished executing all of the GL commands issued before
the end of the previous draw call on the same viewport.

When the new `ready_to_draw' function returns false, drawing of this
widget (on X11) is deferred, spinning the event and worker update loop
until the widget is ready to draw.

This fixes a bug on some GL implementations where many frames of draw
commands may queue up, introducing a large input delay. This occurs when
draw commands are very GPU-expensive, such as the current zone octree
rendering.