Add basic python scripting code
This change introduces a new CMake cache option 'FEATURE_SCRIPTING'
which, when ON (the default), enables linking with the Python 3 library
for future scripting purposes.

The README and build option docs have been updated for this change.

The 'gui' subcommand now initializes the scripting engine during launch
A new flag '--disable-scripting' has been added to skip initialization.
This is important because Python init failure results in a Fatal Error
which terminates the program.