scripting: Add multithreading support
This change allows python threads to execute. Prior to this change, all
python threads would indefinitely block except when executing python
code.

This change introduces recursive 'lock' and 'unlock' functions which are
used to globally lock the entire scripting engine.

Because some types (especially res::atom and res::ref<T>) are not
currently thread safe despite being mutable outside of a transaction,
this commit includes a hack to always take the global lock. The lock is
released during gui blocking calls while waiting for events (select on
X11, MsgWaitForMultipleObjectsEx for WINAPI). In the future, when these
types are properly thread-safe, this hack should be removed.