Start experimental scripting docs
This is a single-page script which is intended to be buildable with
Sphinx, and hopefully human-readable in plain text form as well. No
sphinx configuration file is included at this time.

I have not yet decided if Sphinx is appropriate for this use case, so
this document is placed into a new 'expr' directory instead of 'docs'.