Add travis-ci workaround for cmake/pyenv issues
The error message looks like this:

    CMake Error at /usr/local/cmake-3.12.4/share/cmake-3.12/Modules/FindPackageHandleStandardArgs.cmake:137 (message):

      Could NOT find Python3 (missing: Python3_EXECUTABLE Interpreter)

    Call Stack (most recent call first):

      /usr/local/cmake-3.12.4/share/cmake-3.12/Modules/FindPackageHandleStandardArgs.cmake:378 (_FPHSA_FAILURE_MESSAGE)

      /usr/local/cmake-3.12.4/share/cmake-3.12/Modules/FindPython/Support.cmake:795 (find_package_handle_standard_args)

      /usr/local/cmake-3.12.4/share/cmake-3.12/Modules/FindPython3.cmake:138 (include)

      CMakeLists.txt:324 (find_package)

This change fixes this issue by (almost) nuking the pyenv setup on
travis. It has not been tested if this issue will also occur on other
systems using pyenv, however on those systems it should be possible to
set 'Python3_ROOT_DIR' in the CMake cache to the appropriate python
version directory, e.g. '$(pyenv root)/versions/3.6.7'.