Improve DRNSF_DEP_ARCH detection
This change fixes DRNSF_DEP_ARCH autodetection when using newer versions
of CMake which provide a new "platform" setting for selecting the target
architecture for a given generator. In prior versions, a generator was
tied to a specific architecture, and the older DRNSF_DEP_ARCH detection
mechanism was built on this.