edit::asset_tree: Fix missing on_select event
This fixes a bug where a change in the asset_tree's base atom (such as
when a project change occurs) causes the selected node to be cleared but
fails to raise the on_select event to report the selection change.

This bug may lead to a crash in the following scenario:

  1. Open an NSF file.

  2. Select an asset in the Asset mode.

  3. Close the project.

The selected asset remains in the view despite the asset not actually
existing anymore as the project was destroyed.